Start modulizing the package
