+++ klyqa_ctl/devices/vacuum.py
@@ -1,0 +1,163 @@
+"""Vacuum cleaner"""
+from __future__ import annotations
+from device import *
+
+## Vacuum Cleaner ##
+
+VC_WORKINGSTATUS: Type[Enum] = Enum(
+    "VC_WORKINGSTATUS",
+    "SLEEP STANDBY CLEANING CLEANING_AUTO CLEANING_RANDOM CLEANING_SROOM CLEANING_EDGE CLEANING_SPOT CLEANING_COMP DOCKING CHARGING CHARGING_DC CHARGING_COMP ERROR",
+)
+
+VC_SUCTION_STRENGTHS: Type[Enum] = Enum(
+    "VC_SUCTION_STRENGTHS",
+    "NULL STRONG SMALL NORMAL MAX",
+)
+
+VC_WORKINGMODE: Type[Enum] = Enum(
+    "VC_WORKINGMODE",
+    "STANDBY RANDOM SMART WALL_FOLLOW MOP SPIRAL PARTIAL_BOW SROOM CHARGE_GO"
+)
+
+class KlyqaVCResponseStatus(KlyqaDeviceResponse):
+    """KlyqaVCResponseStatus"""
+
+# Decrypted:  b'{"type":"statechange","mcu":"online","power":"on",
+# "cleaning":"on","beeping":"off","battery":57,"sidebrush":10,
+# "rollingbrush":30,"filter":60,"carpetbooster":200,"area":999,
+# "time":999,"calibrationtime":19999999,"workingmode":null,
+# "workingstatus":"STANDBY","suction":"MID","water":"LOW","direction":"STOP",
+# "errors":["COLLISION","GROUND_CHECK","LEFT_WHEEL","RIGHT_WHEEL","SIDE_SCAN","MID_SWEEP","FAN","TRASH","BATTERY","ISSUES"],
+# "cleaningrec":[],"equipmentmodel":"","alarmmessages":"","commissioninfo":"","action":"get"}
+
+    def __str__(self) -> str:
+        """__str__"""
+        return get_obj_attr_values_as_string(self)
+
+    def __init__(
+        self,
+        **kwargs,
+    ) -> None:
+        """__init__"""
+        self.action: str = ""
+        self.active_command: int = -1
+        self.alarmmessages: str = ""
+        self.area: int = -1
+        self.beeping: str = ""
+        self.battery: str = ""
+        self.calibrationtime: int = -1
+        self.carpetbooster: int = -1
+        self.cleaning: str = ""
+        self.cleaningrec: list[str] = []
+        self.connected: bool = False
+        self.commissioninfo: str = ""
+        self.direction: str = ""
+        self.errors: list[str] = []
+        self.equipmentmodel: str = ""
+        self.filter: int = -1
+        self.filter_tresh: int = -1
+        self.fwversion: str = ""
+        self.id: str = ""
+        self.lastActivityTime: str = ""
+        self.map_parameter: str = ""
+        self.mcuversion: str = ""
+        self.open_slots: int = -1
+        self.power: str = ""
+        self.rollingbrush_tresh: int = -1
+        self.rollingbrush: int = -1
+        self.sdkversion: str = ""
+        self.sidebrush: str = ""
+        self.sidebrush_tresh: int = -1
+        self.suction: str = ""
+        self.time: int = -1
+        self.ts: datetime.datetime = datetime.datetime.now()
+        self.watertank: str = ""
+        self.workingmode: str | None = None
+        self.workingstatus: str = ""
+
+        LOGGER.debug(f"save status {self}")
+        super().__init__(**kwargs)
+
+    def update(self, **kwargs) -> None:
+        super().update(**kwargs)
+
+
+class KlyqaVC(KlyqaDevice):
+    """Klyqa vaccum cleaner"""
+
+    # status: KlyqaVCResponseStatus = None
+
+    def __init__(self) -> None:
+        super().__init__()
+        # self.status = KlyqaVCResponseStatus()
+        self.response_classes["status"] = KlyqaVCResponseStatus
+        
+        
+def add_command_args_cleaner(parser: argparse.ArgumentParser) -> None:
+
+    sub = parser.add_subparsers(title="subcommands", dest='command')
+
+    ota = sub.add_parser('ota', help='allows over the air programming of the device')
+    ota.add_argument('url', help='specify http URL for ota')
+
+    ping = sub.add_parser('ping', help='send a ping and nothing else')
+
+    frs = sub.add_parser('factory-reset', help='trigger a factory reset on the device - the device has to be onboarded again afterwards)')
+
+    reb = sub.add_parser('reboot', help='trigger a reboot')
+
+    prd = sub.add_parser('productinfo', help='get product information')
+
+    req = sub.add_parser('get', help='send state request')
+    req.add_argument('--all', help='If this flag is set, the whole state will be requested', action='store_true')
+    req.add_argument('--power', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--cleaning', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--beeping', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--battery', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--sidebrush', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--rollingbrush', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--filter', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--carpetbooster', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--area', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--time', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--calibrationtime', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--workingmode', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--workingstatus', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--suction', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--water', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--direction', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--errors', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--cleaningrec', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--equipmentmodel', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--alarmmessages', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--commissioninfo', help='If this flag is set, the state element will be requested', action='store_true')
+    req.add_argument('--mcu', help='Ask if mcu is online', action='store_true')
+
+    #device specific
+    set_parser = sub.add_parser("set", help='enables use of the vc1 control arguments and will control vc1')
+    set_parser.add_argument('--power', choices=['on', 'off'], help='turn power on/off')
+    set_parser.add_argument('--cleaning', choices=['on', 'off'], help='turn cleaning on/off')
+    set_parser.add_argument('--beeping', choices=['on', 'off'], help='enable/disable the find-vc function')
+    set_parser.add_argument('--carpetbooster', metavar='strength', type=int, help='set the carpet booster strength (0-255)')
+    set_parser.add_argument('--workingmode', choices=[m.name for m in VC_WORKINGMODE], help='set the working mode')
+    set_parser.add_argument('--water', choices=['LOW', 'MID', 'HIGH'], help='set water quantity')
+    set_parser.add_argument('--suction', choices=[m.name for m in VC_SUCTION_STRENGTHS], help='set suction power')
+    set_parser.add_argument('--direction', choices=['FORWARDS','BACKWARDS', 'TURN_LEFT', 'TURN_RIGHT', 'STOP'], help='manually control movement')
+    set_parser.add_argument('--commissioninfo', type=str, help='set up to 256 characters of commisioning info')
+    set_parser.add_argument('--calibrationtime', metavar='time', type=int, help='set the calibration time (1-1999999999)')
+
+    reset_parser = sub.add_parser("reset", help='enables resetting consumables')
+    reset_parser.add_argument('--sidebrush', help='resets the sidebrush life counter', action='store_true')
+    reset_parser.add_argument('--rollingbrush', help='resets the rollingbrush life counter', action='store_true')
+    reset_parser.add_argument('--filter', help='resets the filter life counter', action='store_true')
+
+    routine_parser = sub.add_parser("routine", help='routine functions')
+    routine_parser.add_argument('--list', help='lists stored routines', action='store_const', const=True, default=False)
+    routine_parser.add_argument('--put', help='store new routine', action='store_const', const=True, default=False)
+    routine_parser.add_argument('--delete', help='delete routine', action='store_const', const=True, default=False)
+    routine_parser.add_argument('--start', help='start routine', action='store_const', const=True, default=False)
+    routine_parser.add_argument('--id', help='specify routine id to act on (for put, start, delete)')
+    routine_parser.add_argument('--scene', help='specify routine scene label (for put)')
+    routine_parser.add_argument('--count', help='get the current free slots for routines', type=bool)
+    routine_parser.add_argument('--commands', help='specify routine program (for put)')
+    # routine_parser.set_defaults(func=routine_request)
