+++ klyqa_ctl/klyqa_ctl.py
@@ -33,6 +33,7 @@ import select
 import logging
 from typing import TypeVar, Any, Type
 
+
 NoneType: Type[None] = type(None)
 import requests, uuid, json
 import os.path
@@ -47,638 +48,32 @@ from collections.abc import Callable
 
 import slugify as unicode_slug
 
+from devices.device import *
+from devices.light import *
+from devices.vacuum import *
+from general.message import *
+from general.general import *
+from general.connections import *
+
 try:
     from Cryptodome.Cipher import AES  # provided by pycryptodome
-    from Cryptodome.Random import get_random_bytes  # pycryptodome
 except:
     from Crypto.Cipher import AES  # provided by pycryptodome
-    from Crypto.Random import get_random_bytes  # pycryptodome
 
 from typing import TypeVar
 
-# DEFAULT_SEND_TIMEOUT_MS=5000000000
-DEFAULT_SEND_TIMEOUT_MS = 500000
-
-LOGGER: logging.Logger = logging.getLogger(__package__)
-LOGGER.setLevel(level=logging.INFO)
-formatter: logging.Formatter = logging.Formatter("%(asctime)s %(levelname)-8s - %(message)s")
-
-logging_hdl = logging.StreamHandler()
-logging_hdl.setLevel(level=logging.INFO)
-logging_hdl.setFormatter(formatter)
-
-LOGGER.addHandler(logging_hdl)
 
 s: str = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S,%f")[:-3]
 
 print(f"{s} start")
 
-TypeJSON = dict[str, Any]
-
-""" string output separator width """
-sep_width = 0
-
-AES_KEY_DEV: bytes = bytes(
-    [
-        0x00,
-        0x11,
-        0x22,
-        0x33,
-        0x44,
-        0x55,
-        0x66,
-        0x77,
-        0x88,
-        0x99,
-        0xAA,
-        0xBB,
-        0xCC,
-        0xDD,
-        0xEE,
-        0xFF,
-    ]
-)
-
-## Bulbs ##
-
-BULB_SCENES: list[dict[str, Any]] = [
-    {
-        "id": 100,
-        "colors": ["#FFECD8", "#FFAA5B"],
-        "label": "Warm White",
-        "commands": "5ch 0 0 0 0 65535 65535 500;p 1000;",
-        "cwww": True,
-    },
-    {
-        "id": 101,
-        "colors": ["#FCF4DD", "#FED07A"],
-        "label": "Daylight",
-        "commands": "5ch 0 0 0 24903 40631 65535 500;p 1000;",
-    },
-    {
-        "id": 102,
-        "colors": ["#FFFFFF", "#B6DAFF"],
-        "label": "Cold White",
-        "commands": "5ch 0 0 0 65535 0 65535 500;p 1000;",
-        "cwww": True,
-    },
-    {
-        "id": 103,
-        "colors": ["#E06004", "#55230D"],
-        "label": "Night Light",
-        "commands": "5ch 9830 3276 0 1310 0 65535 500;p 1000;",
-        "cwww": True,
-    },
-    {
-        "id": 106,
-        "colors": ["#FDCB78", "#FCDA60"],
-        "label": "Relax",
-        "commands": "5ch 26214 26214 0 0 13107 65535 500;p 1000;",
-    },
-    {
-        "id": 109,
-        "colors": ["#090064", "#2E2A5A"],
-        "label": "TV Time",
-        "commands": "5ch 655 0 6553 1310 0 65535 500;p 1000;",
-    },
-    {
-        "id": 107,
-        "colors": ["#FFD1A2", "#FFEDDA"],
-        "label": "Comfort",
-        "commands": "5ch 47185 0 0 18349 0 65535 500;p 1000;",
-    },
-    {
-        "id": 108,
-        "colors": ["#EAFCFF", "#81DFF0"],
-        "label": "Focused",
-        "commands": "5ch 0 0 26214 5000 0 65535 500;p 1000;",
-        "cwww": True,
-    },
-    {
-        "id": 110,
-        "colors": ["#CD3700", "#CD0000", "#CD6600"],
-        "label": "Fireplace",
-        "commands": "5ch 32767 0 0 0 1500 65535 500;p 500;5ch 55535 200 0 0 2000 65535 500;p 500;",
-        # '5ch 32767 0 0 0 1500 65535 500;p 500;5ch 55535 200 0 0 2000 65535 500;p 500;5ch 32767 30000 0 0 1500 65535 500;p 500;5ch 25535 20000 0 0 2000 65535 500;p 500;5ch 62767 0 0 0 1500 65535 500;p 500;5ch 65535 200 0 0 2000 65535 500;p 500;5ch 32767 30000 0 0 1500 65535 500;p 500;5ch 25535 30000 0 0 2000 65535 500;p 500;',
-    },
-    {
-        "id": 122,
-        "colors": ["#12126C", "#B22222", "#D02090"],
-        "label": "Jazz Club",
-        "commands": "5ch 45746 8738 8738 0 0 65535 6100;p 5100;5ch 45232 12336 24672 0 0 65535 6100;p 5000;5ch 53436 8224 37008 0 0 65535 6100;p 5000;5ch 0 0 33896 0 0 65535 6000;p 5000;5ch 18504 15677 35728 0 0 65535 6100;p 5000;5ch 38036 0 52428 0 0 65535 6600;p 5000;5ch 45232 12336 24672 0 0 65535 6100;p 5000;",
-    },
-    {
-        "id": 104,
-        "colors": ["#B20C26", "#CC1933", "#EE6AA7"],
-        "label": "Romantic",
-        "commands": "5ch 58981 6553 16383 0 0 65535 1000;p 7400;5ch 45874 3276 9830 0 0 65535 4400;p 6600;5ch 52428 6553 13107 0 0 65535 8800;p 15200;5ch 41287 0 0 0 0 65535 4400;p 13200;",
-    },
-    {
-        "id": 112,
-        "colors": ["#FFDCE8", "#D2FFD2", "#CCFFFF"],
-        "label": "Gentle",
-        "commands": "5ch 51117 0 0 13107 0 65535 26000;p 56000;5ch 26214 26214 0 8519 0 65535 26000;p 56000;5ch 0 51117 0 13107 0 65535 26000;p 56000;5ch 0 26214 26214 8519 0 65535 26000;p 56000;5ch 0 0 51117 13107 0 65535 26000;p 56000;5ch 26214 0 26214 8519 0 65535 26000;p 56000;",
-    },
-    {
-        "id": 113,
-        "colors": ["#EEAD0E", "#FF7F24", "#CD0000"],
-        "label": "Summer",
-        "commands": "5ch 17039 25558 0 13762 0 65535 8000;p 14000;5ch 39321 7864 0 15728 0 65535 8000;p 14000;5ch 28180 17694 0 11140 0 65535 8000;p 14000;",
-    },
-    {
-        "id": 114,
-        "colors": ["#00BA0C", "#008400", "#0C4400"],
-        "label": "Jungle",
-        "commands": "5ch 0 47840 3276 0 0 65535 2600;p 2100;5ch 5898 10485 1310 0 0 65535 2300;p 4200;5ch 0 34078 0 0 0 65535 2100;p 2500;5ch 3276 17694 0 0 0 65535 4600;p 4000;5ch 9174 46529 0 0 0 65535 5500;p 6900;5ch 9830 43908 1966 0 0 65535 2700;p 4700;5ch 0 55704 0 0 0 65535 2000;p 3800;",
-    },
-    {
-        "id": 105,
-        "colors": ["#00008B", "#0000FF", "#1874ED"],
-        "label": "Ocean",
-        "commands": "5ch 1310 17694 36044 0 0 65535 2400;p 5400;5ch 655 15073 39321 0 0 65535 2100;p 5100;5ch 1310 36044 17039 0 0 65535 4200;p 5100;5ch 1966 22281 29490 0 0 65535 2800;p 5700;5ch 655 19005 34733 0 0 65535 2100;p 4900;5ch 655 26869 27524 0 0 65535 2600;p 3400;5ch 655 26869 27524 0 0 65535 2700;p 3600;5ch 1310 38010 15728 0 0 65535 4200;p 5000;",
-    },
-    # {
-    #   "id": 111,
-    #   "colors": ['#A31900', '#A52300', '#B71933', '#A5237F', '#B71900'],
-    #   "label": 'Club',
-    #   "commands":
-    #     '5ch 41942 6553 0 1310 0 65535 600;p 800;5ch 42597 9174 0 1310 0 65535 8700;p 12000;5ch 47185 6553 13107 1310 0 65535 8700;p 12000;5ch 42597 9174 32767 1310 0 65535 300;p 400;5ch 47185 6553 0 1310 0 65535 300;p 1300;',
-    # },
-    {
-        "id": 115,
-        "colors": ["#EE4000", "#CD6600", "#FFA500"],
-        "label": "Fall",
-        "commands": "5ch 49151 1966 0 9830 0 65535 8400;p 8610;5ch 35388 13107 0 6553 0 65535 8400;p 8750;5ch 52428 0 0 10485 0 65535 8400;p 8740;5ch 39321 9174 0 12451 0 65535 500;p 840;",
-    },
-    {
-        "id": 116,
-        "colors": ["#FFF0F5", "#FF6EB4", "#FF4500"],
-        "label": "Sunset",
-        "commands": "5ch 39321 0 15073 2621 0 65535 5680;p 5880;5ch 51117 0 0 13107 0 65535 5680;p 5880;5ch 43253 11796 0 2621 0 65535 5680;p 5880;5ch 38010 0 15073 7208 0 65535 5680;p 5880;5ch 46529 0 0 3932 0 65535 5680;p 5880;5ch 41287 11140 0 7864 0 65535 5680;p 5880;",
-    },
-    {
-        "id": 117,
-        "colors": ["#FF0000", "#0000FF", "#00FF00"],
-        "label": "Party",
-        "commands": "5ch 55704 0 0 0 0 65535 132;p 272;5ch 55704 0 0 0 0 65535 132;p 272;5ch 0 55704 0 0 0 65535 132;p 272;5ch 0 55704 0 0 0 65535 132;p 272;5ch 0 0 55704 0 0 65535 132;p 272;5ch 0 0 55704 0 0 65535 132;p 272;5ch 28180 0 27524 0 0 65535 132;p 272;5ch 0 28180 27524 0 0 65535 132;p 272;",
-    },
-    {
-        "id": 118,
-        "colors": ["#F0FFF0", "#C1FFC1", "#FFE4E1"],
-        "label": "Spring",
-        "commands": "5ch 19660 15728 19660 0 0 65535 8000;p 11000;5ch 20315 26214 13107 0 0 65535 8000;p 11000;5ch 17039 19005 19005 0 0 65535 8000;p 11000;5ch 20315 14417 14417 0 0 65535 8000;p 11000;5ch 19005 18349 17694 0 0 65535 8000;p 11000;5ch 11796 30146 6553 0 0 65535 8000;p 11000;",
-    },
-    {
-        "id": 119,
-        "colors": ["#C1FFC1", "#C0FF3E", "#CAFF70"],
-        "label": "Forest",
-        "commands": "5ch 23592 22937 0 3932 0 65535 6000;p 8000;5ch 19005 23592 0 7864 0 65535 6200;p 10100;5ch 22281 21626 0 12451 0 65535 6000;p 10000;5ch 23592 22281 0 4587 0 65535 5800;p 10400;5ch 18349 27524 0 1966 0 65535 6200;p 7000;5ch 8519 25558 0 23592 0 65535 6200;p 9400;",
-    },
-    {
-        "id": 120,
-        "colors": ["#104E8B", "#00008B", "#4876FF"],
-        "label": "Deep Sea",
-        "commands": "5ch 3932 3276 59636 0 0 65535 4100;p 5100;5ch 3276 6553 53738 0 0 65535 4100;p 5000;5ch 0 0 43908 0 0 65535 4100;p 5000;5ch 655 1310 53083 0 0 65535 3600;p 5000;5ch 1310 0 53738 0 0 65535 4000;p 5000;",
-    },
-    {
-        "id": 121,
-        "colors": ["#90ee90", "#8DEEEE", "#008B45"],
-        "label": "Tropical",
-        "commands": "5ch 0 43253 0 0 36044 65535 3000;p 4000;5ch 0 0 0 0 65535 65535 2400;p 5400;5ch 0 38010 0 0 48495 65535 2600;p 3600;5ch 0 32767 0 0 0 65535 2000;p 3400;5ch 0 46529 0 0 26869 65535 3100;p 4100;5ch 0 43908 0 0 0 65535 4000;p 7000;5ch 0 49806 0 0 16383 65535 2000;p 5000;",
-    },
-    {
-        "id": 123,
-        "colors": ["#FF6AD0", "#8BFFC7", "#96A0FF"],
-        "label": "Magic Mood",
-        "commands": "5ch 65535 27242 53456 0 0 35535 2400;p 1180;5ch 30326 33924 65535 0 0 35535 2200;p 1110;5ch 65535 21331 21331 0 0 35535 2800;p 1200;5ch 35723 55535 31143 0 0 35535 2800;p 1200;5ch 38550 41120 65535 0 0 35535 2400;p 1040;5ch 65535 61423 29041 0 0 35535 2400;p 1000;",
-    },
-    {
-        "id": 124,
-        "colors": ["#FF0000", "#B953FF", "#DBFF96"],
-        "label": "Mystic Mountain",
-        "commands": "5ch 65535 0 0 0 0 35535 1400;p 980;5ch 65535 30326 52685 0 0 35535 1200;p 910;5ch 47543 21331 65535 0 0 35535 1800;p 1200;5ch 35723 65535 44461 0 0 35535 1800;p 1200;5ch 56283 65535 38550 0 0 35535 1400;p 1040;5ch 65535 29041 53456 0 0 35535 1400;p 1000;",
-    },
-    {
-        "id": 125,
-        "colors": ["#FB0000", "#FFF748", "#B97FFF"],
-        "label": "Cotton Candy",
-        "commands": "5ch 65535 0 52428 0 0 35535 1400;p 980;5ch 47545 32639 655350 0 0 35535 1200;p 910;5ch 65535 33410 33410 0 0 35535 1800;p 1200;5ch 65535 63479 18504 0 0 35535 1800;p 1200;5ch 65535 63222 16448 0 0 35535 1400;p 1040;5ch 64507 0 0 0 0 35535 1400;p 1000;",
-    },
-    {
-        "id": 126,
-        "colors": ["#8BFFE1", "#D8FA97", "#FF927F"],
-        "label": "Ice Cream",
-        "commands": "5ch 65535 0 0 0 0 35535 1400;p 980;5ch 65535 37522 32639 0 0 35535 1200;p 910;5ch 61166 54741 65535 0 0 35535 1800;p 1200;5ch 35723 65535 57825 0 0 35535 1800;p 1200;5ch 55512 64250 38807 0 0 35535 1400;p 1040;5ch 65535 56796 62709 0 0 35535 1400;p 1000;",
-    },
-]
-
-## Vacuum Cleaner ##
-
-VC_WORKINGSTATUS: Type[Enum] = Enum(
-    "VC_WORKINGSTATUS",
-    "SLEEP STANDBY CLEANING CLEANING_AUTO CLEANING_RANDOM CLEANING_SROOM CLEANING_EDGE CLEANING_SPOT CLEANING_COMP DOCKING CHARGING CHARGING_DC CHARGING_COMP ERROR",
-)
-
-VC_SUCTION_STRENGTHS: Type[Enum] = Enum(
-    "VC_SUCTION_STRENGTHS",
-    "NULL STRONG SMALL NORMAL MAX",
-)
-
-VC_WORKINGMODE: Type[Enum] = Enum(
-    "VC_WORKINGMODE",
-    "STANDBY RANDOM SMART WALL_FOLLOW MOP SPIRAL PARTIAL_BOW SROOM CHARGE_GO"
-)
-
-class AsyncIOLock:
-    """AsyncIOLock"""
-
-    task: asyncio.Task
-    lock: asyncio.Lock
-    _instance = None
-
-    def __init__(self) -> None:
-        """__init__"""
-        self.lock = asyncio.Lock()
-        self.task = None
-
-    async def acquire(self) -> None:
-        """acquire"""
-        await self.lock.acquire()
-        self.task = asyncio.current_task()
-
-    def release(self) -> None:
-        """release"""
-        self.lock.release()
-
-    def force_unlock(self) -> bool:
-        """force_unlock"""
-        try:
-            self.task.cancel()
-            self.lock.release()
-        except:
-            return False
-        return True
-
-    @classmethod
-    def instance(
-        cls,
-    ):
-        """instance"""
-        if cls._instance is None:
-            LOGGER.debug("Creating new AsyncIOLock instance")
-            cls._instance = cls.__new__(cls)
-            # Put any initialization here.
-            cls._instance.__init__()
-        return cls._instance
-
 
 tcp_udp_port_lock: AsyncIOLock = AsyncIOLock.instance()
-
-
-class RefParse:
-    """RefParse"""
-
-    ref = None
-
-    def __init__(self, ref) -> None:
-        self.ref = ref
-
-
-async def async_json_cache(json_data, json_file) -> tuple [Device_config, bool]:
-    """
-    If json data is given write it to cache json_file.
-    Else try to read from json_file the cache.
-    """
-
-    return_json: Device_config = json_data
-    cached = False
-    if json_data:
-        """
-        save the json for offline cache in dirpath where called script resides
-        ids: { sets }
-        """
-        return_json = json_data
-        try:
-            s = ""
-            for id, sets in json_data.items():
-                if isinstance(sets, (datetime.datetime, datetime.date)):
-                    sets = sets.isoformat()
-                s = s + '"' + id + '": ' + json.dumps(sets) + ", "
-            s = "{" + s[:-2] + "}"
-            async with aiofiles.open(
-                os.path.dirname(sys.argv[0]) + f"/{json_file}", mode="w"
-            ) as f:
-                await f.write(s)
-        except Exception as e:
-            LOGGER.warning(f'Could not save cache for json file "{json_file}".')
-    else:
-        """no json data, take cached json from disk if available"""
-        try:
-            async with aiofiles.open(
-                os.path.dirname(sys.argv[0]) + f"/{json_file}", mode="r"
-            ) as f:
-                s = await f.read()
-            return_json = json.loads(s)
-            cached = True
-        except:
-            LOGGER.warning(f'No cache from json file "{json_file}" available.')
-    return (return_json, cached)
-
-
-def get_fields(object):
-    """get_fields"""
-    if hasattr(object, "__dict__"):
-        return object.__dict__.keys()
-    else:
-        return dir(object)
-
-
-def get_obj_attrs_as_string(object) -> str:
-    """get_obj_attrs_as_string"""
-    fields = get_fields(object)
-    attrs = [
-        a for a in fields if not a.startswith("__") and not callable(getattr(object, a))
-    ]
-    return ", ".join(attrs)
-
-
-def get_obj_attr_values_as_string(object) -> str:
-    """get_obj_attr_values_as_string"""
-    fields = get_fields(object)
-    attrs = [
-        a for a in fields if not a.startswith("__") and not callable(getattr(object, a))
-    ]
-    vals = []
-    for a in attrs:
-        _str: str = str(getattr(object, a))
-        vals.append(_str if _str else '""')
-    return ", ".join(vals)
-
-
-class RGBColor:
-    """RGBColor"""
-
-    r: int
-    g: int
-    b: int
-
-    def __str__(self) -> str:
-        return "[" + ",".join([str(self.r), str(self.g), str(self.b)]) + "]"
-        # return get_obj_values_as_string(self)
-
-    def __init__(self, r: int, g: int, b: int) -> None:
-        self.r = r
-        self.g = g
-        self.b = b
-
-
-def format_uid(text: str) -> str:
-    return unicode_slug.slugify(text)
-
-class KlyqaDeviceResponse:
-    # type: str = ""
-
-    def __init__(
-        self,
-        # type: str,
-        **kwargs) -> None:
-        """__init__"""
-        self.type: str = ""
-        self.ts: datetime.datetime = datetime.datetime.now()
-        self.update(**kwargs)
-
-    def update(self, **kwargs) -> None:
-        # Walk through parsed kwargs dict and look if names in dict exists as attribute in class,
-        # then apply the value in kwargs to the value in class.
-        for attr in kwargs:
-            if hasattr(self, attr):
-                setattr(self, attr, kwargs[attr])
-
-# class KlyqaDeviceResponse:
-#     """KlyqaDeviceResponse"""
-
-#     type: str = ""
-#     response_msg: dict = {}
-#     ts: datetime.datetime = datetime.datetime.now()
-
-#     def __init__(self, type=type, response_msg: dict = {}) -> None:
-#         self.type = type
-#         self.response_msg = response_msg
-#         self.ts = datetime.datetime.now()
-
-
-# eventually dataclass
-class KlyqaDeviceResponseIdent(KlyqaDeviceResponse):
-    """KlyqaDeviceResponseIdent"""
-
-    def __init__(
-        self,
-        **kwargs,
-    ) -> None:
-        self.fw_version: str = ""
-        self.fw_build: str = ""
-        self.hw_version: str = ""
-        self.manufacturer_id: str = ""
-        self.product_id: str = ""
-        self.sdk_version: str = ""
-        self._unit_id: str = ""
-        super().__init__(**kwargs)
-
-    def update(self, **kwargs) -> None:
-        super().update(**kwargs)
-
-    @property
-    def unit_id(self) -> str:
-        return self._unit_id
-
-    @unit_id.setter
-    def unit_id(self, unit_id: str) -> None:
-        self._unit_id = format_uid(unit_id)
-
-class KlyqaVCResponseStatus(KlyqaDeviceResponse):
-    """KlyqaVCResponseStatus"""
-
-# Decrypted:  b'{"type":"statechange","mcu":"online","power":"on",
-# "cleaning":"on","beeping":"off","battery":57,"sidebrush":10,
-# "rollingbrush":30,"filter":60,"carpetbooster":200,"area":999,
-# "time":999,"calibrationtime":19999999,"workingmode":null,
-# "workingstatus":"STANDBY","suction":"MID","water":"LOW","direction":"STOP",
-# "errors":["COLLISION","GROUND_CHECK","LEFT_WHEEL","RIGHT_WHEEL","SIDE_SCAN","MID_SWEEP","FAN","TRASH","BATTERY","ISSUES"],
-# "cleaningrec":[],"equipmentmodel":"","alarmmessages":"","commissioninfo":"","action":"get"}
-
-    def __str__(self) -> str:
-        """__str__"""
-        return get_obj_attr_values_as_string(self)
-
-    def __init__(
-        self,
-        **kwargs,
-    ) -> None:
-        """__init__"""
-        self.action: str = ""
-        self.active_command: int = -1
-        self.alarmmessages: str = ""
-        self.area: int = -1
-        self.beeping: str = ""
-        self.battery: str = ""
-        self.calibrationtime: int = -1
-        self.carpetbooster: int = -1
-        self.cleaning: str = ""
-        self.cleaningrec: list[str] = []
-        self.connected: bool = False
-        self.commissioninfo: str = ""
-        self.direction: str = ""
-        self.errors: list[str] = []
-        self.equipmentmodel: str = ""
-        self.filter: int = -1
-        self.filter_tresh: int = -1
-        self.fwversion: str = ""
-        self.id: str = ""
-        self.lastActivityTime: str = ""
-        self.map_parameter: str = ""
-        self.mcuversion: str = ""
-        self.open_slots: int = -1
-        self.power: str = ""
-        self.rollingbrush_tresh: int = -1
-        self.rollingbrush: int = -1
-        self.sdkversion: str = ""
-        self.sidebrush: str = ""
-        self.sidebrush_tresh: int = -1
-        self.suction: str = ""
-        self.time: int = -1
-        self.ts: datetime.datetime = datetime.datetime.now()
-        self.watertank: str = ""
-        self.workingmode: str | None = None
-        self.workingstatus: str = ""
-
-        LOGGER.debug(f"save status {self}")
-        super().__init__(**kwargs)
-
-    def update(self, **kwargs) -> None:
-        super().update(**kwargs)
-
-class KlyqaBulbResponseStatus(KlyqaDeviceResponse):
-    """Klyqa_Bulb_Response_Status"""
-
-    # status: str = ""
-    # color: RGBColor = RGBColor(-1, -1, -1)
-    # brightness: int = -1
-    # temperature: int = -1
-    # active_command: int = -1
-    # active_scene: str = ""
-    # open_slots: int
-    # mode: str = ""
-    # fwversion: str = ""
-    # sdkversion: str = ""
-    # ts: datetime.datetime = datetime.datetime.now()
-
-    def __str__(self) -> str:
-        """__str__"""
-        return get_obj_attr_values_as_string(self)
-
-    def __init__(
-        self,
-        # type: str,
-        # status: str = "off",
-        # brightness: dict[str, int] = {},
-        # mode: str = "",
-        # open_slots: int = 0,
-        # fwversion: str = "",
-        # sdkversion: str = "",
-        # color: dict[str, int] = {},
-        # temperature: int = -1,
-        # active_command: int = 0,
-        # active_scene: str = 0,
-        **kwargs
-    ) -> None:
-        """__init__"""
-        # self.status = status
-        # self.color = (
-        #     RGBColor(color["red"], color["green"], color["blue"]) if color else {}
-        # )
-        # self.brightness = brightness["percentage"] if brightness else 0
-        # self.temperature = temperature
-        # self.active_command = active_command
-        # self.active_scene = active_scene
-        # self.open_slots = open_slots
-        # self.mode = mode
-        # self.fwversion = fwversion
-        # self.sdkversion = sdkversion
-        self.active_command: int | None = None
-        self.active_scene: str | None = None
-        self._brightness: int | None = None
-        self._color: RGBColor | None = None
-        self.fwversion: str | None = None
-        self.mode: str | None = None # cmd, cct, rgb
-        self.open_slots: int | None = None
-        self.sdkversion: str | None = None
-        self.status: str | None = None
-        self.temperature: int | None = None
-        LOGGER.debug(f"save status {self}")
-        super().__init__(**kwargs)
-
-    @property
-    def brightness(self) -> int | None:
-        return self._brightness
-
-    @brightness.setter
-    def brightness(self, brightness: dict[str, int]) -> None:
-        self._brightness = int(brightness["percentage"])
-
-    @property
-    def color(self) -> RGBColor | None:
-        return self._color
-
-    @color.setter
-    def color(self, color: dict[str, int]) -> None:
-        self._color = RGBColor(color["red"], color["green"], color["blue"]) if color else None
-
-
-Device_config = dict
 
 # Device profiles for limits and features (traits) for the devices.
 #
 device_configs: dict[str, Device_config] = dict()
 
-class LocalConnection:
-    """LocalConnection"""
-
-    state: str = "WAIT_IV"
-    localIv: bytes = get_random_bytes(8)
-    remoteIv: bytes = b""
-
-    sendingAES: bytes = None
-    receivingAES: bytes = None
-    address: dict[str, str | int] = {"ip": "", "port": -1}
-    socket: socket.socket = None
-    received_packages = []
-    sent_msg_answer = {}
-    aes_key_confirmed: bool = False
-
-
-    def __init__(self) -> None:
-        self.state = "WAIT_IV"
-        self.localIv = get_random_bytes(8)
-
-        self.sendingAES = None
-        self.receivingAES = None
-        self.address = {"ip": "", "port": -1}
-        self.socket: socket.socket = None
-        self.received_packages = []
-        self.sent_msg_answer = {}
-        self.aes_key_confirmed = False
-        self.started: datetime = datetime.datetime.now()
-
-
-class CloudConnection:
-    """CloudConnection"""
-
-    received_packages = []
-    connected: bool
-
-    def __init__(self) -> None:
-        self.connected = False
-        self.received_packages = []
 
 
 Device_TCP_return: Type[Device_TCP_return] = Enum(
@@ -687,353 +82,11 @@ Device_TCP_return: Type[Device_TCP_return] = Enum(
 )
 
 
-Message_state: Type[Message_state] = Enum("Message_state", "sent answered unsent")
-
-
-@dataclass
-class Message:
-    started: datetime.datetime
-    msg_queue: list[tuple]
-    msg_queue_sent = []  #: list[str] = dataclasses.field(default_factory=list)
-    args: list[str]
-    target_uid: str
-    state: Message_state = Message_state.unsent
-    finished: datetime.datetime = None
-    answer: str = ""
-    answer_utf8: str = ""
-    answer_json = {}
-    # callback on error event or answer
-    callback: Callable[[Any], None] = NoneType
-    time_to_live_secs: int = -1
-    msg_counter: int = -1
-
-    def __post_init__(self) -> None:
-        # super().__init__(self, *args, **kwargs)
-        global MSG_COUNTER
-        self.msg_counter = MSG_COUNTER
-        MSG_COUNTER = MSG_COUNTER + 1
-
-    async def call_cb(self) -> None:
-        await self.callback(self, self.target_uid)
-
-    async def check_msg_ttl(self) -> bool:
-        if datetime.datetime.now() - self.started > datetime.timedelta(
-            seconds=self.time_to_live_secs
-        ):
-            LOGGER.debug(
-                f"time to live {self.time_to_live_secs} seconds for message {self.msg_counter} {self.msg_queue} ended."
-            )
-            if self.callback:
-                await self.call_cb()
-            return False
-        return True
-
-class KlyqaDevice:
-    """KlyqaDevice"""
-
-    local_addr: dict[str, Any] = {"ip": "", "port": -1}
-    cloud: CloudConnection
-
-    u_id: str = "no_uid"
-    acc_sets: dict = {}
-    """ account settings """
-
-    _use_lock: asyncio.Lock
-    _use_thread: asyncio.Task
-
-    recv_msg_unproc: list[Message]
-    ident: KlyqaDeviceResponseIdent = None
-
-    response_classes = {}
-    status: KlyqaDeviceResponse | None
-
-    def __init__(self) -> None:
-        self.local_addr = {"ip": "", "port": -1}
-        self.cloud = CloudConnection()
-        self.ident: KlyqaDeviceResponseIdent = KlyqaDeviceResponseIdent()
-
-        self.u_id: str = "no_uid"
-        self.acc_sets = {}
-        self._use_lock = None
-        self._use_thread = None
-        self.recv_msg_unproc = []
-
-        self.status = None
-        self.response_classes = {
-            "ident": KlyqaDeviceResponseIdent,
-            "status": KlyqaDeviceResponse,
-        }
-
-    def process_msgs(self) -> None:
-        for msg in self.recv_msg_unproc:
-            LOGGER.debug(f"updating device {self.u_id} entity with msg:")
-            self.recv_msg_unproc.remove(msg)
-
-    def get_name(self) -> str:
-        return (
-            f"{self.acc_sets['name']} ({self.u_id})"
-            if self.acc_sets and "name" in self.acc_sets and self.acc_sets
-            else self.u_id
-        )
-
-    async def use_lock(self, timeout=30, **kwargs) -> bool:
-        try:
-            if not self._use_lock:
-                self._use_lock = asyncio.Lock()
-
-            LOGGER.debug(f"wait for lock... {self.get_name()}")
-
-            if await self._use_lock.acquire():
-                self._use_thread = asyncio.current_task()
-                LOGGER.debug(f"got lock... {self.get_name()}")
-                return True
-        except asyncio.TimeoutError:
-            LOGGER.error(f'Timeout for getting the lock for device "{self.get_name()}"')
-        except Exception as excp:
-            LOGGER.debug(f"different error while trying to lock.")
-
-        return False
-
-    async def use_unlock(self) -> None:
-        if not self._use_lock:
-            self._use_lock = asyncio.Lock()
-        if self._use_lock.locked() and self._use_thread == asyncio.current_task():
-            try:
-                self._use_lock.release()
-                self._use_thread = None
-                LOGGER.debug(f"got unlock... {self.get_name()}")
-            except:
-                pass
-
-    def save_device_message(self, msg) -> None:
-        """msg: json dict"""
-
-        status_update_types: set={"status", "statechange"}
-        if msg["type"] in status_update_types:
-            msg["type"] = "status"
-        if "type" in msg and hasattr(self, msg["type"]): #and msg["type"] in msg:
-            try:
-                LOGGER.debug(f"save device msg {msg} {self.ident} {self.u_id}")
-                if msg["type"] == "ident":
-                    # setattr(self, "ident", self.ident.update(**msg))
-                    # setattr(
-                    #     self,
-                    #     msg["type"],
-                    #     self.response_classes[msg["type"]](**msg[msg["type"]]),
-                    # )
-                    self.ident.update(**msg["ident"])
-                elif msg["type"] in status_update_types: # and self.status:
-                    # setattr(self, msg["type"], self.response_classes[msg["type"]](**msg))
-                    # setattr(self, "status", self.status.update(**msg))
-                    if self.status is None:
-                        self.status = self.response_classes["status"](**msg)
-                    else:
-                        self.status.update(**msg)
-            except Exception as e:
-                LOGGER.error(f"{traceback.format_exc()}")
-                LOGGER.error("Could not process device response: ")
-                LOGGER.error(str(msg))
-
-class KlyqaVC(KlyqaDevice):
-    """Klyqa vaccum cleaner"""
-
-    # status: KlyqaVCResponseStatus = None
-
-    def __init__(self) -> None:
-        super().__init__()
-        # self.status = KlyqaVCResponseStatus()
-        self.response_classes["status"] = KlyqaVCResponseStatus
-
-class KlyqaBulb(KlyqaDevice):
-    """KlyqaBulb"""
-
-    # status: KlyqaBulbResponseStatus = None
-
-    def __init__(self) -> None:
-        super().__init__()
-        # self.status = KlyqaBulbResponseStatus()
-        self.response_classes["status"] = KlyqaBulbResponseStatus
-
-    def setTemp(self, temp: int):
-        temperature_enum = []
-        try:
-            temperature_enum = [
-                trait["value_schema"]["properties"]["colorTemperature"]["enum"]
-                for trait in device_configs[self.ident.product_id]["deviceTraits"]
-                if trait["trait"] == "@core/traits/color-temperature"
-            ]
-            if len(temperature_enum) < 2:
-                raise Exception()
-        except:
-            LOGGER.error("No temperature change on the bulb available")
-            return False
-        if temp < temperature_enum[0] or temp > temperature_enum[1]:
-            LOGGER.error(
-                "Temperature for bulb out of range ["
-                + temperature_enum[0]
-                + ", "
-                + temperature_enum[1]
-                + "]."
-            )
-            return False
-
-
 ReturnTuple = TypeVar("ReturnTuple", tuple[int, str], tuple[int, dict])
 
 
-def send_msg(msg, device: KlyqaDevice, connection: LocalConnection):
-    info_str: str = (
-        'Sending in local network to "'
-        + device.get_name()
-        + '": '
-        + json.dumps(json.loads(msg), sort_keys=True, indent=4)
-    )
-
-    LOGGER.info(info_str)
-    plain = msg.encode("utf-8")
-    while len(plain) % 16:
-        plain = plain + bytes([0x20])
-
-    cipher = connection.sendingAES.encrypt(plain)
-
-    while True:
-        try:
-            connection.socket.send(
-                bytes([len(cipher) // 256, len(cipher) % 256, 0, 2]) + cipher
-            )
-            return True
-        except socket.timeout:
-            LOGGER.debug("Send timed out, retrying...")
-            pass
-    return False
-
-
-def color_message(red, green, blue, transition, skipWait=False) -> tuple[str, int]:
-    waitTime = transition if not skipWait else 0
-    return (
-        json.dumps(
-            {
-                "type": "request",
-                "color": {
-                    "red": red,
-                    "green": green,
-                    "blue": blue,
-                },
-                "transitionTime": transition,
-            }
-        ),
-        waitTime,
-    )
-
-
-def temperature_message(temperature, transition, skipWait=False) -> tuple[str, int]:
-    waitTime = transition if not skipWait else 0
-    return (
-        json.dumps(
-            {
-                "type": "request",
-                "temperature": temperature,
-                "transitionTime": transition,
-            }
-        ),
-        waitTime,
-    )
-
-
-def percent_color_message(red, green, blue, warm, cold, transition, skipWait) -> tuple[str, int]:
-    waitTime = transition if not skipWait else 0
-    return (
-        json.dumps(
-            {
-                "type": "request",
-                "p_color": {
-                    "red": red,
-                    "green": green,
-                    "blue": blue,
-                    "warm": warm,
-                    "cold": cold,
-                    # "brightness" : brightness
-                },
-                "transitionTime": transition,
-            }
-        ),
-        waitTime,
-    )
-
-
-def brightness_message(brightness, transition) -> tuple[str, int]:
-    return (
-        json.dumps(
-            {
-                "type": "request",
-                "brightness": {
-                    "percentage": brightness,
-                },
-                "transitionTime": transition,
-            }
-        ),
-        transition,
-    )
-
 
 AES_KEYs: dict[str, bytes] = {}
-
-
-PRODUCT_URLS: dict[str, str] = {
-    """TODO: Make permalinks for urls."""
-    "@klyqa.lighting.cw-ww.gu10": "https://klyqa.de/produkte/gu10-white-strahler",
-    "@klyqa.lighting.rgb-cw-ww.gu10": "https://klyqa.de/produkte/gu10-color-strahler",
-    "@klyqa.lighting.cw-ww.g95": "https://www.klyqa.de/produkte/g95-vintage-lampe",
-    "@klyqa.lighting.rgb-cw-ww.e14": "https://klyqa.de/produkte/e14-color-lampe",
-    "@klyqa.lighting.cw-ww.e14": "https://klyqa.de/produkte/gu10-white-strahler",
-    "@klyqa.lighting.rgb-cw-ww.e27": "https://www.klyqa.de/produkte/e27-color-lampe",
-    "@klyqa.lighting.cw-ww.e27": "https://klyqa.de/produkte/e27-white-lampe",
-    "@klyqa.cleaning.vc1": "https://klyqa.de/Alle-Produkte/Smarter-Starter",
-}
-
-commands_send_to_bulb: list[str] = [
-    "request",
-    "ping",
-    "color",
-    "temperature",
-    "brightness",
-    "routine_list",
-    "routine_put",
-    "routine_id",
-    "routine_commands",
-    "routine_delete",
-    "routine_start",
-    "power",
-    "reboot",
-    "factory_reset",
-    "WW",
-    "daylight",
-    "CW",
-    "nightlight",
-    "relax",
-    "TVtime",
-    "comfort",
-    "focused",
-    "fireplace",
-    "club",
-    "romantic",
-    "gentle",
-    "summer",
-    "jungle",
-    "ocean",
-    "fall",
-    "sunset",
-    "party",
-    "spring",
-    "forest",
-    "deep_sea",
-    "tropical",
-    "magic",
-    "mystic",
-    "cotton",
-    "ice",
-    "command"
-]
 
 S = TypeVar("S", argparse.ArgumentParser, type(None))
 
@@ -1200,279 +253,6 @@ def add_config_args(parser: argparse.ArgumentParser) -> None:
         const=True,
         default=False,
     )
-
-
-def add_command_args_bulb(parser: argparse.ArgumentParser) -> None:
-    """Add arguments to the argument parser object.
-
-    Args:
-        parser: Parser object
-    """
-    parser.add_argument(
-        "--transitionTime", nargs=1, help="transition time in milliseconds", default=[0]
-    )
-
-    parser.add_argument("--color", nargs=3, help="set color command (r,g,b) 0-255")
-    parser.add_argument(
-        "--temperature",
-        nargs=1,
-        help="set temperature command (kelvin range depends on lamp profile) (lower: warm, higher: cold)",
-    )
-    parser.add_argument("--brightness", nargs=1, help="set brightness in percent 0-100")
-    parser.add_argument(
-        "--percent_color",
-        nargs=5,
-        metavar=("RED", "GREEN", "BLUE", "WARM", "COLD"),
-        help="set colors and white tones in percent 0 - 100",
-    )
-    parser.add_argument("--ota", nargs=1, help="specify http URL for ota")
-    parser.add_argument(
-        "--ping", help="send ping", action="store_const", const=True, default=False
-    )
-    parser.add_argument(
-        "--request",
-        help="send status request",
-        action="store_const",
-        const=True,
-        default=False,
-    )
-
-    # TODO: currently not fully implemented
-    # parser.add_argument(
-    #     "--identity",
-    #     help="send status request",
-    #     action="store_const",
-    #     const=True,
-    #     default=False,
-    # )
-
-    parser.add_argument(
-        "--factory_reset",
-        help="trigger a factory reset on the device (Warning: device has to be onboarded again afterwards)",
-        action="store_const",
-        const=True,
-        default=False,
-    )
-    parser.add_argument(
-        "--routine_list",
-        help="lists stored routines",
-        action="store_const",
-        const=True,
-        default=False,
-    )
-    parser.add_argument(
-        "--routine_put",
-        help="store new routine",
-        action="store_const",
-        const=True,
-        default=False,
-    )
-    parser.add_argument(
-        "--routine_delete",
-        help="delete routine",
-        action="store_const",
-        const=True,
-        default=False,
-    )
-    parser.add_argument(
-        "--routine_start",
-        help="start routine",
-        action="store_const",
-        const=True,
-        default=False,
-    )
-    parser.add_argument(
-        "--routine_id", help="specify routine id to act on (for put, start, delete)"
-    )
-    parser.add_argument("--routine_scene", help="specify routine scene label (for put)")
-    parser.add_argument("--routine_commands", help="specify routine program (for put)")
-    parser.add_argument(
-        "--reboot",
-        help="trigger a reboot",
-        action="store_const",
-        const=True,
-        default=False,
-    )
-    parser.add_argument(
-        "--power", nargs=1, metavar='"on"/"off"', help="turns the bulb on/off"
-    )
-    parser.add_argument(
-        "--enable_tb", nargs=1, help="enable thingsboard connection (yes/no)"
-    )
-
-    parser.add_argument(
-        "--fade", nargs=2, help="fade in/out time in milliseconds on powering device on/off", metavar=("IN", "OUT")
-    )
-
-    # parser.add_argument("--loop", help="loop", action="store_true")
-    parser.add_argument("--WW", help="Warm White", action="store_true")
-    parser.add_argument("--daylight", help="daylight", action="store_true")
-    parser.add_argument("--CW", help="Cold White", action="store_true")
-    parser.add_argument("--nightlight", help="nightlight", action="store_true")
-    parser.add_argument("--relax", help="relax", action="store_true")
-    parser.add_argument("--TVtime", help="TVtime", action="store_true")
-    parser.add_argument("--comfort", help="comfort", action="store_true")
-    parser.add_argument("--focused", help="focused", action="store_true")
-    parser.add_argument("--fireplace", help="fireplace", action="store_true")
-    parser.add_argument("--club", help="club", action="store_true")
-    parser.add_argument("--romantic", help="romantic", action="store_true")
-    parser.add_argument("--gentle", help="gentle", action="store_true")
-    parser.add_argument("--summer", help="summer", action="store_true")
-    parser.add_argument("--jungle", help="jungle", action="store_true")
-    parser.add_argument("--ocean", help="ocean", action="store_true")
-    parser.add_argument("--fall", help="fall", action="store_true")
-    parser.add_argument("--sunset", help="sunset", action="store_true")
-    parser.add_argument("--party", help="party", action="store_true")
-    parser.add_argument("--spring", help="spring", action="store_true")
-    parser.add_argument("--forest", help="forest", action="store_true")
-    parser.add_argument("--deep_sea", help="deep_sea", action="store_true")
-    parser.add_argument("--tropical", help="tropical", action="store_true")
-    parser.add_argument("--magic", help="magic Mood", action="store_true")
-    parser.add_argument("--mystic", help="Mystic Mountain", action="store_true")
-    parser.add_argument("--cotton", help="Cotton Candy", action="store_true")
-    parser.add_argument("--ice", help="Ice Cream", action="store_true")
-
-
-def add_command_args_cleaner(parser: argparse.ArgumentParser) -> None:
-
-    sub = parser.add_subparsers(title="subcommands", dest='command')
-
-    ota = sub.add_parser('ota', help='allows over the air programming of the device')
-    ota.add_argument('url', help='specify http URL for ota')
-
-    ping = sub.add_parser('ping', help='send a ping and nothing else')
-
-    frs = sub.add_parser('factory-reset', help='trigger a factory reset on the device - the device has to be onboarded again afterwards)')
-
-    reb = sub.add_parser('reboot', help='trigger a reboot')
-
-    prd = sub.add_parser('productinfo', help='get product information')
-
-    req = sub.add_parser('get', help='send state request')
-    req.add_argument('--all', help='If this flag is set, the whole state will be requested', action='store_true')
-    req.add_argument('--power', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--cleaning', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--beeping', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--battery', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--sidebrush', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--rollingbrush', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--filter', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--carpetbooster', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--area', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--time', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--calibrationtime', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--workingmode', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--workingstatus', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--suction', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--water', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--direction', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--errors', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--cleaningrec', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--equipmentmodel', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--alarmmessages', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--commissioninfo', help='If this flag is set, the state element will be requested', action='store_true')
-    req.add_argument('--mcu', help='Ask if mcu is online', action='store_true')
-
-    #device specific
-    set_parser = sub.add_parser("set", help='enables use of the vc1 control arguments and will control vc1')
-    set_parser.add_argument('--power', choices=['on', 'off'], help='turn power on/off')
-    set_parser.add_argument('--cleaning', choices=['on', 'off'], help='turn cleaning on/off')
-    set_parser.add_argument('--beeping', choices=['on', 'off'], help='enable/disable the find-vc function')
-    set_parser.add_argument('--carpetbooster', metavar='strength', type=int, help='set the carpet booster strength (0-255)')
-    set_parser.add_argument('--workingmode', choices=[m.name for m in VC_WORKINGMODE], help='set the working mode')
-    set_parser.add_argument('--water', choices=['LOW', 'MID', 'HIGH'], help='set water quantity')
-    set_parser.add_argument('--suction', choices=[m.name for m in VC_SUCTION_STRENGTHS], help='set suction power')
-    set_parser.add_argument('--direction', choices=['FORWARDS','BACKWARDS', 'TURN_LEFT', 'TURN_RIGHT', 'STOP'], help='manually control movement')
-    set_parser.add_argument('--commissioninfo', type=str, help='set up to 256 characters of commisioning info')
-    set_parser.add_argument('--calibrationtime', metavar='time', type=int, help='set the calibration time (1-1999999999)')
-
-    reset_parser = sub.add_parser("reset", help='enables resetting consumables')
-    reset_parser.add_argument('--sidebrush', help='resets the sidebrush life counter', action='store_true')
-    reset_parser.add_argument('--rollingbrush', help='resets the rollingbrush life counter', action='store_true')
-    reset_parser.add_argument('--filter', help='resets the filter life counter', action='store_true')
-
-    routine_parser = sub.add_parser("routine", help='routine functions')
-    routine_parser.add_argument('--list', help='lists stored routines', action='store_const', const=True, default=False)
-    routine_parser.add_argument('--put', help='store new routine', action='store_const', const=True, default=False)
-    routine_parser.add_argument('--delete', help='delete routine', action='store_const', const=True, default=False)
-    routine_parser.add_argument('--start', help='start routine', action='store_const', const=True, default=False)
-    routine_parser.add_argument('--id', help='specify routine id to act on (for put, start, delete)')
-    routine_parser.add_argument('--scene', help='specify routine scene label (for put)')
-    routine_parser.add_argument('--count', help='get the current free slots for routines', type=bool)
-    routine_parser.add_argument('--commands', help='specify routine program (for put)')
-    # routine_parser.set_defaults(func=routine_request)
-
-
-PROD_HOST = "https://app-api.prod.qconnex.io"
-TEST_HOST = "https://app-api.test.qconnex.io"
-
-
-SEND_LOOP_MAX_SLEEP_TIME = 0.05
-MSG_COUNTER = 0
-
-
-class Data_communicator:
-    def __init__(self, server_ip: str = "0.0.0.0") -> None:
-        self.tcp: socket.socket = None
-        self.udp: socket.socket = None
-        self.server_ip: str = server_ip
-
-    def shutdown(self) -> None:
-
-        try:
-            if self.tcp:
-                self.tcp.shutdown(socket.SHUT_RDWR)
-                self.tcp.close()
-                LOGGER.debug("Closed TCP port 3333")
-                self.tcp = None
-        except:
-            pass
-
-        try:
-            if self.udp:
-                self.udp.close()
-                LOGGER.debug("Closed UDP port 2222")
-                self.udp = None
-        except:
-            pass
-
-    async def bind_ports(self) -> bool:
-        """bind ports."""
-        # await tcp_udp_port_lock.acquire()
-        self.shutdown()
-        server_address: tuple[str, int]
-        try:
-
-            self.udp = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-            self.udp.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
-            self.udp.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-            if self.server_ip is not None:
-                server_address = (self.server_ip, 2222)
-            else:
-                server_address = ("0.0.0.0", 2222)
-            self.udp.bind(server_address)
-            LOGGER.debug("Bound UDP port 2222")
-
-        except Exception as e:
-            LOGGER.error(
-                "Error on opening and binding the udp port 2222 on host for initiating the device communication."
-            )
-            return False
-
-        try:
-            self.tcp = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            self.tcp.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-            server_address = ("0.0.0.0", 3333)
-            self.tcp.bind(server_address)
-            LOGGER.debug("Bound TCP port 3333")
-            self.tcp.listen(1)
-
-        except Exception as e:
-            LOGGER.error(
-                "Error on opening and binding the tcp port 3333 on host for initiating the device communication."
-            )
-            return False
-        return True
 
 
 class Klyqa_account:
