+++ test/test_send_msg.py
@@ -1,0 +1,2 @@
+import klyqa_ctl
+
